mod: 반환 값 정보 변경
